+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 
-### Unreleased (BREAKING)
+### Unreleased
+
+## v2.1.0 (2021-03-08)
+
+* Require that the config values for session hash salt, garbage collect token and tokenista secret
+  are defined (not null) before the dependency container creates the related service.
+* Add extensible HTTP-triggered garbage collection (see README)
+* Add a dependency factory for our mysql session handler
 
 ## v2.0.2 (2020-11-02)
 * Support SwiftMailer ^6